requirements.txt
asgiref==3.8.1
black==24.4.2
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
Django==5.0.7
django-environ==0.11.2
djangorestframework==3.15.2
drf-writable-nested==0.7.0
filelock==3.15.4
identify==2.6.0
iniconfig==2.0.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.8.0
pytest==8.3.2
pytest-django==4.8.0
PyYAML==6.0.1
qrcode==8.0
snowballstemmer==2.2.0
sqlparse==0.5.1
virtualenv==20.26.3
